Move participants controller to api/v1
[#139486493]
